requirements.txt
dj-database-url==0.5.0
Django==2.1.8
django-bootstrap-modal-forms==1.4.2
django-crispy-forms==1.7.2
django-environ==0.4.5
django-model-utils==3.1.2
django-tinymce4-lite==1.7.5
django-widget-tweaks==1.4.3
factory-boy==2.12.0
Faker==1.0.7
gunicorn==19.9.0
jsmin==2.2.2
lorem==0.1.1
Pillow==6.0.0
psycopg2==2.8.2
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
transliterate==1.10.2
whitenoise==4.1.2
